# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/articles/{articleId}/assets" : {
      "get" : {
        "tags" : [ "article-asset-controller" ],
        "operationId" : "list_1",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ArticleAssetResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "article-asset-controller" ],
        "operationId" : "upload",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "file" : {
                    "type" : "string",
                    "format" : "binary"
                  }
                },
                "required" : [ "file" ]
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArticleAssetResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/articles/{articleId}/assets/{assetId}" : {
      "get" : {
        "tags" : [ "article-asset-controller" ],
        "operationId" : "content",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "assetId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string",
                  "format" : "byte"
                }
              }
            }
          }
        }
      }
    },
    "/articles/{articleId}/assets/{assetId}" : {
      "get" : {
        "tags" : [ "article-asset-controller" ],
        "operationId" : "content_1",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "assetId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string",
                  "format" : "byte"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ArticleAssetResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "articleId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "originalFilename" : {
            "type" : "string"
          },
          "contentType" : {
            "type" : "string"
          },
          "byteSize" : {
            "type" : "integer",
            "format" : "int64"
          },
          "markdownUrl" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
    }
  }
}